refac: malloc_init, alloc
