+++ Makefile
@@ -5,8 +5,9 @@
 SRCS	:= malloc.c \
 			show_alloc_heap.c \
 			show_free_list.c \
-			page_size.c \
-			align.c \
+			malloc_init.c \
+			alloc.c \
+			page_size.c align.c \
 			free.c \
 			freelst.c
 #			realloc.c \
