+++ includes/alloc.h
@@ -1,0 +1,8 @@
+#ifndef ALLOC_H 
+#define ALLOC_H 
+
+#include <stddef.h>
+
+void	*alloc_pages_by_size(size_t map_size, void *start);
+
+#endif /* ALLOC_H */
