+++ includes/malloc.h
@@ -71,6 +71,7 @@ void	show_alloc_heap();
 void	show_free_list(t_region region);
 void	add_chunk_to_freelist(t_malloc_chunk *chunk, t_malloc_chunk **freelist);
 
+void	init_malloc();
 
 void	freelst_add_front(t_malloc_chunk **lst, t_malloc_chunk *new_chunk);
 void	freelst_insert(t_malloc_chunk *prev, t_malloc_chunk *new_chunk);
